==> machinehack/ode_to_code_2022/notebooks/1-baseline.ipynb <==
# Imports and Data overview

train_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/train.csv'
test_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/test.csv'
sub_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/sample_submission.csv'

!pip install --upgrade --quiet optuna

!pip install --quiet xgboost==1.5

import time 

import numpy as np
from scipy.stats import mode
import pandas as pd
pd.set_option("precision", 4)

from sklearn.preprocessing import LabelEncoder
from sklearn.feature_selection import chi2, f_classif, mutual_info_classif 
from sklearn.metrics import roc_auc_score, accuracy_score
from sklearn.model_selection import train_test_split, StratifiedKFold

import optuna
from optuna.samplers import TPESampler
from optuna.integration import XGBoostPruningCallback
from xgboost import XGBClassifier

SEED = 2311

train = pd.read_csv(train_url)
test = pd.read_csv(test_url)

train.head()

train.info()

test.info()

test.drop(['season'], axis=1, inplace=True)

# Missing values

train.isna().sum()

test.isna().sum()

train['gill-attachment'].unique(), test['gill-attachment'].unique()

train['ring-type'].unique(), test['ring-type'].unique()

According to the data description, 'f' corresponds to 'none' for both these features with missing values.  

Thus, we will fill missing values with 'f'.

train.fillna('f', axis=1, inplace=True)
test.fillna('f', axis=1, inplace=True)

# EDA and Feature Engineering

train['season'].value_counts()

features = [f for f in train.columns if f not in ('season', 'fold')]
num_features = ['cap-diameter', 'stem-height', 'stem-width']
cat_features = [f for f in features if f not in num_features]
cat_mask = [f in cat_features for f in features]

train[cat_features].describe()

train[num_features].describe()

train[cat_features] = train[cat_features].astype('category')
test[cat_features] = test[cat_features].astype('category')

target_labels = LabelEncoder()
train['season'] = target_labels.fit_transform(train['season'])

# Hyperparameter tuning

base_params = {
    'objective': 'multi:softmax',
    'booster': 'gbtree',
    'use_label_encoder': False,
    'tree_method': 'gpu_hist',
    'predictor': 'gpu_predictor',
    'enable_categorical': True,
    'n_estimators': 5000,
    'eval_metric': 'mlogloss',
    'random_state': SEED,
    'verbosity': 0
}

def objective(trial, xtrain, ytrain, xval, yval, base_params):
  param_grid = {
    'learning_rate': trial.suggest_float('learning_rate', 0.01, 0.3),
    'subsample': trial.suggest_float('subsample', 0.5, 1.0),
    'reg_lambda': trial.suggest_loguniform('reg_lambda', 1e-8, 1.0),
    'reg_alpha': trial.suggest_loguniform('reg_alpha', 1e-8, 1.0),
    'max_depth': trial.suggest_int('max_depth', 3, 12)
  }

  model = XGBClassifier(**base_params, **param_grid)
    
  model.fit(
    xtrain, ytrain,
    eval_set=[(xval, yval)],
    early_stopping_rounds=100,
    callbacks=[
      XGBoostPruningCallback(trial, 'validation_0-mlogloss')
    ],
    verbose=100
  )
    
  val_predictions = model.predict(xval)
    
  return accuracy_score(yval, val_predictions)

xtrain, xval, ytrain, yval = train_test_split(train[features], train['season'],
                                              test_size=0.2,
                                              stratify=train['season'],
                                              shuffle=True,
                                              random_state=SEED)

study = optuna.create_study(direction='minimize', 
                            sampler=TPESampler(), 
                            study_name='MH_ode_to_code')

%%time
study.optimize(
    lambda trial: objective(trial, xtrain, ytrain, xval, yval, base_params),
    n_trials=50
)

print(f'Best value (Accuracy): {study.best_value:.5f}')

best_params = study.best_params
print('Best params:')
for key, value in best_params.items():
    print(f'\t{key}: {value}')

# Cross-validation + Inference

N_SPLITS = 5
train['fold'] = -1

skf = StratifiedKFold(n_splits=N_SPLITS, shuffle=True, random_state=SEED)

for fold, (_, val_idx) in enumerate(skf.split(X=train, y=train['season'])):
  train.loc[val_idx, 'fold'] = fold

def custom_cross_val_predict(train, test, features, model):
  oof_preds = {}
  test_preds = []
  scores = []
    
  cv_start = time.time()
    
  for fold in range(N_SPLITS):
    print('-' * 40)
        
    xtrain = train[train.fold != fold].reset_index(drop=True)
    ytrain = xtrain.season

    xval = train[train.fold == fold].reset_index(drop=True)
    yval = xval.season
    val_idx = xval.index.tolist()
    
    fold_start = time.time()
        
    model.fit(
      xtrain[features], ytrain,
      eval_set=[(xval[features], yval)],
      early_stopping_rounds=100,
      verbose=100
    )
    
    val_preds = model.predict(xval[features])
    oof_preds.update(dict(zip(val_idx, val_preds)))
    acc = accuracy_score(yval, val_preds)
    scores.append(acc)
        
    fold_end = time.time()
        
    print(f'Fold #{fold}: Accuracy = {acc:.5f} \
    [Time: {fold_end - fold_start:.2f}s]')
        
    test_preds.append(model.predict(test[features]))
        
    cv_end = time.time()
  
  print(f'Average accuracy = {np.mean(scores):.5f} \
  with std. dev. = {np.std(scores):.5f}')
  
  print(f'[Total time: {cv_end - cv_start:.2f}s]')
    
  oof_preds = pd.DataFrame.from_dict(oof_preds, orient='index').reset_index()
  test_preds = mode(np.column_stack(test_preds), axis=1).mode
    
  return oof_preds, test_preds

model = XGBClassifier(**base_params, **best_params)

oof_preds, test_preds = custom_cross_val_predict(train, test, features, model)

sub1 = pd.read_csv(sub_url)
sub1['season'] = target_labels.inverse_transform(test_preds.ravel())
sub1.to_csv('sub1.csv', index=False)

!head sub1.csv
