==> machinehack/ode_to_code_2022/notebooks/2-automl.ipynb <==
train_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/train.csv'
test_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/test.csv'
sub_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/sample_submission.csv'

import gc
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd

train = pd.read_csv(train_url)

test = pd.read_csv(test_url)
test.drop(['season'], axis=1, inplace=True)

sub = pd.read_csv(sub_url)

# AutoGluon (Model 2)

!python3 -m pip install --quiet pip
!python3 -m pip install --quiet setuptools wheel
!python3 -m pip install --quiet "mxnet<2.0.0"
!python3 -m pip install --quiet autogluon

from autogluon.tabular import TabularPredictor

label = 'season'
eval_metric = 'accuracy'
save_path = '/AutoGluons/'

predictor = TabularPredictor(label=label, eval_metric=eval_metric, 
                             path=save_path, verbosity=2)
predictor.fit(train, presets='best_quality', time_limit=600)

results = predictor.fit_summary()

preds = predictor.predict(test)

sub2 = sub.copy()
sub2['season'] = preds
sub2.to_csv('sub2.csv', index=False)

# del predictor
# gc.collect()

# H20 AutoML (Model 3)

!pip install --quiet --upgrade h2o

import h2o
from h2o.automl import H2OAutoML

h2o.init()

h2o_train = h2o.H2OFrame(train)
h2o_test = h2o.H2OFrame(test)

h2o_train['season'] = h2o.H2OFrame(train['season'].to_numpy().ravel()).asfactor()

features = [f for f in test.columns]

model3 = H2OAutoML(stopping_metric='misclassification', max_runtime_secs=600)
model3.train(x=features, y='season', training_frame=h2o_train)

model3.leaderboard

preds3 = model3.leader.predict(h2o_test).as_data_frame()['predict']

sub3 = sub.copy()
sub3['season'] = preds3
sub3.to_csv('sub3.csv', index=False)

# del model3
# gc.collect()

# MLJAR (Model 4)

# !pip install --quiet --uprade mljar-supervised

# from supervised.automl import AutoML
# from sklearn.metrics import accuracy_score

# model4 = AutoML(mode='compete', 
#                 ml_task='multiclass_classification', 
#                 eval_metric='accuracy', 
#                 total_time_limit=600)

# model4.fit(X=train[features], y=train['season'])

# preds4 = model4.predict(test)

# sub4 = sub.copy()
# sub4['season'] = preds4
# sub4.to_csv('sub4.csv', index=False)

# del model4
# gc.collect()

# FLAML (Model 5)

!pip install --quiet --upgrade flaml

from flaml import AutoML

model5 = AutoML()
model5.fit(train[test.columns], train['season'], 
           task='classification', time_budget=600)

preds5 = model5.predict(test)

sub5 = sub.copy()
sub5['season'] = preds5
sub5.to_csv('sub5.csv', index=False)
