==> machinehack/ode_to_code_2022/notebooks/3-preprocessing.ipynb <==
# Imports

train_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/train.csv'
test_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/test.csv'
sub_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/sample_submission.csv'

import time
import gc
gc.enable()

import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
pd.set_option("precision", 4)

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
sns.set(style="darkgrid")

from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold

SEED = 2311

train = pd.read_csv(train_url)
test = pd.read_csv(test_url)

# EDA

train.info()

test.info()

target = ['season']
features = [f for f in train.columns if f not in target]
num_features = ['cap-diameter', 'stem-height', 'stem-width']
cat_features = [f for f in features if f not in num_features]
test.drop(['season'], axis=1, inplace=True)

train[num_features].describe()

train[cat_features].describe()

# Missing values

train.isna().sum()

test.isna().sum()

### ring-type

Ring exists but type is missing:

len(train.loc[(train['has-ring'] == 't') & train['ring-type'].isna()])

len(test.loc[(test['has-ring'] == 't') & test['ring-type'].isna()])

Two options:  
1. Fill all missing *ring-type* with 'f', which corresponds to 'none'  
2. Remove rows with missing *ring-type* when *has-ring* is 't', fill with 'f' when *has-ring* is 'f'. (Not possible on test set)

test['ring-type'].fillna('f', inplace=True)

train_full = train.copy()
train_full['ring-type'].fillna('f', inplace=True)

train_reduced = train.drop(train[(train['has-ring'] == 't') & train['ring-type'].isna()].index)
train_reduced['ring-type'].fillna('f', inplace=True)

### gill-attachment  

*gill-attachment* is 'none' (== 'f') implies that the mushroom/fruit does not have gills. ([Source](https://www.zoology.ubc.ca/~biodiv/mushroom/gill_attachment.html))  

We can confirm that information from our data:

len(train.loc[(train['gill-attachment'] == 'f') & (train['gill-color'] != 'f')])

Checking *gill-color* for missing *gill-attachment* values:

len(train.loc[(train['gill-color'] == 'f') & 
              train['gill-attachment'].isna()])

train.loc[train['gill-attachment'].isna()]['gill-color'].value_counts()

len(test.loc[(test['gill-color'] == 'f') & 
              test['gill-attachment'].isna()])

test.loc[test['gill-attachment'].isna()]['gill-color'].value_counts()

test['gill-attachment'].fillna('m', inplace=True) #m -> missing

train_full['gill-attachment'].fillna('m', inplace=True)

train_reduced.drop(
    train_reduced[train_reduced['gill-attachment'].isna() & 
                  (train_reduced['gill-color'] != 'f')].index, 
    inplace=True
)
train_reduced['gill-attachment'].fillna('m', inplace=True)

# Encoding

target_labels = LabelEncoder()

train_full['season'] = target_labels.fit_transform(train_full['season'])
train_reduced['season'] = target_labels.transform(train_reduced['season'])

train.head()

edible_mapping = {'e': 0, 'p': 1}

caps = ['x', 'o', 'c', 'f', 'b', 's', 'p']
cap_mapping = dict(zip(caps, np.arange(len(caps))))

colors= ['f', 'n', 'g', 'y', 'p', 'w', 'o', 'l', 'r', 'e', 'b', 'k', 'u']
color_mapping = dict(zip(colors, np.arange(len(colors))))

bruise_mapping = {'f': 0, 't': 1}

gills = ['f', 's', 'x', 'a', 'e', 'd', 'p']
gill_mapping = dict(zip(gills, np.arange(len(gills))))
gill_mapping['m'] = -1

hasring_mapping = {'f': 0, 't': 1}

ringtypes = ['f', 'p', 'l', 'z', 'g', 'm', 'e', 'r', 'c', 'y', 's']
ringtype_mapping = dict(zip(ringtypes, np.arange(len(ringtypes))))

habitats = ['d', 'l', 'g', 'w', 'm', 'p', 'h', 'u']
habitat_mapping = dict(zip(habitats, np.arange(len(habitats))))

def encode(column, mapping):
  train_full[column] = train_full[column].map(mapping).fillna(train_full[column])
  train_reduced[column] = train_reduced[column].map(mapping).fillna(train_reduced[column])
  test[column] = test[column].map(mapping).fillna(test[column])

encode('edible-poisonous', edible_mapping)
encode('cap-shape', cap_mapping)
encode('cap-color', color_mapping)
encode('does-bruise-or-bleed', bruise_mapping)
encode('gill-attachment', gill_mapping)
encode('gill-color', color_mapping)
encode('stem-color', color_mapping)
encode('has-ring', hasring_mapping)
encode('ring-type', ringtype_mapping)
encode('habitat', habitat_mapping)

# Folds

N_SPLITS = 5

skf = StratifiedKFold(n_splits=N_SPLITS, shuffle=True, random_state=SEED)

train_full['fold'] = -1
for fold, (_, val_idx) in enumerate(skf.split(X=train_full, y=train_full['season'])):
  train_full.loc[val_idx, 'fold'] = fold

train_reduced.reset_index(inplace=True)
train_reduced['fold'] = -1
for fold, (_, val_idx) in enumerate(skf.split(X=train_reduced, y=train_reduced['season'])):
  train_reduced.loc[val_idx, 'fold'] = fold

# Final

train.shape, train_full.shape, train_reduced.shape

test.shape

train_full.sample(5)

test.sample(5)

train_full.to_csv('train_full.csv', index=False)
train_reduced.to_csv('train_reduced.csv', index=False)
test.to_csv('test.csv', index=False)
