==> machinehack/ode_to_code_2022/notebooks/4-data-modelling-catboost.ipynb <==
# Imports

train_full_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/processed/train_full.csv'
train_reduced_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/processed/train_reduced.csv'
test_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/processed/test.csv'
sub_url = 'https://raw.githubusercontent.com/sidt-ai/data-science-competitions/main/machinehack/ode_to_code_2022/data/raw/sample_submission.csv'

%%capture
!python3 -m pip install --quiet --upgrade optuna
!python3 -m pip install --quiet --upgrade xgboost==1.5
!python3 -m pip install --quiet --upgrade catboost
!python3 -m pip install --quiet --upgrade lightgbm

%%capture
!python3 -m pip install --quiet --upgrade sklearn

import time
import gc
gc.enable()

import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
pd.set_option("precision", 4)

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
sns.set(style="darkgrid")

from scipy.stats import mode
from sklearn.metrics import roc_auc_score, accuracy_score
from sklearn.model_selection import train_test_split

import xgboost as xgb
import lightgbm as lgb
from catboost import CatBoostClassifier

import optuna
from optuna.samplers import TPESampler

SEED = 2311

import torch

if torch.cuda.is_available():
    GPU = True
else:
    GPU = False

train_full = pd.read_csv(train_full_url)
train_reduced = pd.read_csv(train_reduced_url)
test = pd.read_csv(test_url)
sub = pd.read_csv(sub_url)

target = ['season']
features = [f for f in train_full.columns if f not in ('season', 'fold')]
num_features = ['cap-diameter', 'stem-height', 'stem-width']
cat_features = [f for f in features if f not in num_features]

target_mapping = {0: 'a', 1: 's', 2: 'u', 3: 'w'}

# CatBoost

### Model 6 - train_full

cb_base = {
    'loss_function': 'MultiClass',
    'custom_metric': 'F1',
    'eval_metric': 'Accuracy',
    'iterations': 5000,
    'random_seed': SEED,
    'use_best_model': True,
    'classes_count': 4,
    'task_type': 'CPU',
    'cat_features': cat_features
}

if GPU:
  cb_base['task_type'] = 'GPU'

def cb_objective(trial, xtrain, ytrain, xval, yval, cb_base):
  
  cb_grid = {
    'learning_rate' :trial.suggest_loguniform('learning_rate', 0.01, 0.3),
    'depth': trial.suggest_int('depth', 1, 15),
    'l2_leaf_reg' :trial.suggest_loguniform('l2_leaf_reg', 1e-8, 20),
    'random_strength' : trial.suggest_loguniform('random_strength', 1, 50),
    'grow_policy':trial.suggest_categorical(
        'grow_policy', ['Lossguide','SymmetricTree']
    ),
    'max_bin': trial.suggest_int("max_bin", 20, 500),
    'min_data_in_leaf':trial.suggest_int('min_data_in_leaf', 1, 100),
    'bootstrap_type': trial.suggest_categorical(
        'bootstrap_type', ['Bayesian', 'Bernoulli']
    )
  }
              
  if cb_grid['grow_policy'] == 'Lossguide':
    cb_grid['max_leaves']:trial.suggest_int('max_leaves', 1, 100)
    
  if cb_grid["bootstrap_type"] == "Bayesian":
    cb_grid["bagging_temperature"] = trial.suggest_float("bagging_temperature", 0, 10)
  elif cb_grid["bootstrap_type"] == "Bernoulli":
    cb_grid["subsample"] = trial.suggest_float("subsample", 0.1, 1)

  model = CatBoostClassifier(**cb_base, **cb_grid)

  model.fit(
      xtrain, ytrain,
      eval_set=[(xval, yval)],
      verbose=100,
      early_stopping_rounds=100
  )

  val_predictions = model.predict(xval)
  return accuracy_score(yval, val_predictions)  

cbfull_study = optuna.create_study(direction='maximize', 
                            sampler=TPESampler(), 
                            study_name='MH_O2C_M6')

xtrain, xval, ytrain, yval = train_test_split(train_full[features], 
                                              train_full[target],
                                              test_size=0.2, shuffle=True,
                                              stratify=train_full[target],
                                              random_state=SEED)

# %%capture
cbfull_study.optimize(
    lambda trial: cb_objective(trial, xtrain, ytrain, xval, yval, cb_base),
    n_trials=50
)

print(f'Best value (Accuracy): {cbfull_study.best_value:.5f}')

cbfull_best = cbfull_study.best_params
print('Best params:')
for key, value in cbfull_best.items():
    print(f'\t{key}: {value}')

N_SPLITS = 5

def custom_cross_val_predict(train, test, features, model):
  oof_preds = {}
  test_preds = []
  scores = []
    
  cv_start = time.time()
    
  for fold in range(N_SPLITS):
    print('-' * 40)
        
    xtrain = train[train.fold != fold].reset_index(drop=True)
    ytrain = xtrain.season

    xval = train[train.fold == fold].reset_index(drop=True)
    yval = xval.season
    val_idx = xval.index.tolist()
    
    fold_start = time.time()
        
    model.fit(
      xtrain[features], ytrain,
      eval_set=[(xval[features], yval)],
      early_stopping_rounds=100,
      verbose=100
    )
    
    val_preds = model.predict(xval[features])
    oof_preds.update(dict(zip(val_idx, val_preds)))
    acc = accuracy_score(yval, val_preds)
    scores.append(acc)
        
    fold_end = time.time()
        
    print(f'Fold #{fold}: Accuracy = {acc:.5f} \
    [Time: {fold_end - fold_start:.2f}s]')
        
    test_preds.append(model.predict(test[features]))
        
    cv_end = time.time()
  
  print(f'Average accuracy = {np.mean(scores):.5f} \
  with std. dev. = {np.std(scores):.5f}')
  
  print(f'[Total time: {cv_end - cv_start:.2f}s]')
    
  oof_preds = pd.DataFrame.from_dict(oof_preds, orient='index').reset_index()
  test_preds = mode(np.column_stack(test_preds), axis=1).mode
    
  return oof_preds, test_preds

model6 = CatBoostClassifier(**cb_base, **cbfull_best)

oof_preds6, test_preds6 = custom_cross_val_predict(train_full, test, features, model6)

sub6 = pd.read_csv(sub_url)
sub6['season'] = test_preds6.ravel()
sub6['season'] = sub6['season'].map(target_mapping)
sub6.to_csv('sub6.csv', index=False)

!head sub6.csv

### Model 7 - train_reduced

cbred_study = optuna.create_study(direction='maximize', 
                            sampler=TPESampler(), 
                            study_name='MH_O2C_M7')

xtrain, xval, ytrain, yval = train_test_split(train_reduced[features], 
                                              train_reduced[target],
                                              test_size=0.2, shuffle=True,
                                              stratify=train_reduced[target],
                                              random_state=SEED)

%%capture
cbred_study.optimize(
    lambda trial: cb_objective(trial, xtrain, ytrain, xval, yval, cb_base),
    n_trials=50
)

print(f'Best value (Accuracy): {cbred_study.best_value:.5f}')

cbred_best = cbred_study.best_params
print('Best params:')
for key, value in cbred_best.items():
    print(f'\t{key}: {value}')

model7 = CatBoostClassifier(**cb_base, **cbred_best)

oof_preds7, test_preds7 = custom_cross_val_predict(train_reduced, test, features, model7)

sub7 = pd.read_csv(sub_url)
sub7['season'] = test_preds7.ravel()
sub7['season'] = sub7['season'].map(target_mapping)
sub7.to_csv('sub7.csv', index=False)

!head sub7.csv
